requirements-minimal.txt
# Dependências mínimas essenciais para o PDF Digest
Flask
Werkzeug
docling
pydantic>=2.6.0,<3.0.0
pydantic-settings
redis
requests
psutil
PyYAML
python-json-logger 